Move loadRemixDllAndInitialize to C API. Fix C API PFN_remixapi_dxvk_SetDefaultOutput accepting C++ reference instead of a pointer. C++ remix::Interface wrapper should accept C structs instead C++ wrapper structs -- for compatibility.
See merge request lightspeedrtx/dxvk-remix-nv!731

+++ src/dxvk/rtx_render/rtx_hashing.h
@@ -156,7 +156,7 @@ namespace dxvk {
     }
 
     // Array of hashes, indexed by HashComponent
-    XXH64_hash_t fields[HashComponents::Count];
+    XXH64_hash_t fields[static_cast<uint32_t>(HashComponents::Count)];
     XXH64_hash_t precombined[rules::Total];
   };
